Update logRankTest.R
exp(sumOMinE/sumVarOMinE) is the Peto estimator and is slightly different from what was implemented before. Could you please give this estimate some attribute that notifies the user that this is the Peto estimator? Maybe add the source to R Documentation description of the function, that is: the Statistical Appendix of this paper: Yusuf, S., Peto, R., Lewis, J., Collins, R., & Sleight, P. (1985). Beta blockade during and after myocardial infarction: an overview of the randomized trials. Progress in cardiovascular diseases, 27(5), 335-371.